backend/requirements.txt
asgiref==3.8.1
Django==5.1.3
graphene==3.4.3
graphene-django==3.2.2
graphql-core==3.2.5
graphql-relay==3.2.0
neo4j-driver==5.27.0
promise==2.3
python-dateutil==2.9.0.post0
pytz==2024.2
six==1.16.0
sqlparse==0.5.2
text-unidecode==1.3
typing_extensions==4.12.2
